backend/requirements.txt
django>=5.1.0
djangorestframework>=3.15.0
django-cors-headers>=4.6.0
python-dotenv>=1.0.0

# AI
google-genai>=1.5.0

# Data
pandas>=2.2.0
openpyxl>=3.1.0
numpy>=1.26.0

# Document parsing
PyMuPDF>=1.24.0
python-docx>=1.1.0

# Charting
plotly>=5.24.0
kaleido>=0.2.1

# PPT
python-pptx>=1.0.0

# Utilities
pillow>=10.4.0
tenacity>=9.0.0
pydantic>=2.9.0

# Testing
pytest>=8.3.0
pytest-django>=4.9.0
pytest-mock>=3.14.0
